[v2.4.x] prov/efa/test: Cast the qp gen when doing the comparison
qp gen is a uint8_t, since unit-test is running in a persistent
process, the qp gen will keep increasing per each qp lifecycle.
Therefore, when there are > 255 QP recreated before executing this
test, the first_gen + 1 wil overflow. assert_init_equal will first
cast first_gen + 1 into a int which will not overflow, so it finally
compare a 0 and 256 and cause failure.

This patch fixes this issue by cast first_gen + 1 to uint8_t first,
which will overflow as expected, before doing the comparison.

(cherry picked from commit dfc41d596ca43b639bc2670a9473ce7f4b206554)

+++ prov/efa/test/efa_unit_test_data_path_direct.c
@@ -231,7 +231,7 @@ void test_efa_data_path_direct_qp_gen_increments_across_qps(struct efa_resource 
 	efa_unit_test_resource_construct(resource, FI_EP_RDM, EFA_DIRECT_FABRIC_NAME);
 	base_ep = container_of(resource->ep, struct efa_base_ep, util_ep.ep_fid);
 
-	assert_int_equal(base_ep->qp->data_path_direct_qp.gen, first_gen + 1);
+	assert_int_equal(base_ep->qp->data_path_direct_qp.gen, (uint8_t) (first_gen + 1));
 #else
 	skip();
 #endif
